requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
akshare>=1.16,<2.0
pandas>=2.2,<3.0
anthropic>=0.55,<1.0
python-dotenv>=1.0,<2.0
